Phase_II/backend/requirements.txt
fastapi>=0.109.0
sqlmodel>=0.0.14
psycopg[binary]>=3.1.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.46
aiosqlite>=0.22.1
asyncpg>=0.31.0
httpx>=0.27.0
pyjwt>=2.8.0
cryptography>=41.0.0